+++ CHANGELOG.rst
@@ -9,6 +9,14 @@ Change Log
 0.9.3
 =====
 
+* Added ALLOW_FOR_RCLONE_BASED_S3_TO_S3_COPY feature to types/file.py to specifically
+  and ONLY allow rclone based S3-to-S3 copy. For some reason rclone needs s3:ListBucket;
+  but we limit it to a key prefix set to exactly our upload file destination key.
+
+
+0.9.3
+=====
+
 * Remove dependency on ``awscli``